time

+++ philo.h
@@ -6,7 +6,7 @@
 /*   By: asauvage <asauvage@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/16 11:50:26 by asauvage          #+#    #+#             */
-/*   Updated: 2026/03/11 16:31:48 by asauvage         ###   ########.fr       */
+/*   Updated: 2026/03/12 11:41:02 by asauvage         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_data
 {
 	int				nb_philo;
 	int				iterate;
-	long long		start_time;
+	long			start_time;
 	int				time_die;
 	int				time_eat;
 	int				time_sleep;
@@ -39,7 +39,7 @@ typedef struct s_data
 typedef struct s_philo
 {
 	int				id;
-	long long		last_meal;
+	long			last_meal;
 	int				nb_meal;
 	pthread_t		thread;
 	pthread_mutex_t	*l_fork;
@@ -56,6 +56,6 @@ void		ft_putstr_fd(char *s, int fd);
 void		clear_all(t_data *data, t_philo *philo);
 void		create_philo(t_data *data, t_philo *philo);
 void		ft_usleep(long long time_in_ms);
-long long	get_time_ms(void);
+long		get_time_ms(void);
 
 #endif
