+++ src/ft_usleep.c
@@ -6,14 +6,14 @@
 /*   By: asauvage <asauvage@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/11 15:25:54 by asauvage          #+#    #+#             */
-/*   Updated: 2026/03/11 16:40:01 by asauvage         ###   ########.fr       */
+/*   Updated: 2026/03/12 11:41:27 by asauvage         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
 
-long long	get_time_ms(void)
+long	get_time_ms(void)
 {
 	struct	timeval	time;
 
@@ -23,7 +23,7 @@ long long	get_time_ms(void)
 
 void	ft_usleep(long long time_in_ms)
 {
-	long long	start_time;
+	long	start_time;
 
 	start_time = get_time_ms();
 	while (get_time_ms() - start_time < time_in_ms)
